Fixed compare of sample title strings (accents, symbols etc.). Refactored rule view compare operator chooser and its resized into base. Refactored table view for sample table and palette table. Added comments inside each method.

+++ Saempl/Source/SampleFileFilterRuleLength.h
@@ -22,8 +22,8 @@ public:
     
     // Methods
     bool matches(SampleItem const & inSampleItem) override;
-    double getCompareValue();
-    void setCompareValue(double const & inCompareValue);
+    int getCompareValue();
+    void setCompareValue(int const & inCompareValue);
     
 private:
     JUCE_HEAVYWEIGHT_LEAK_DETECTOR(SampleFileFilterRuleLength)
